refactor: 로그인 코드 리팩토링
