+++ src/app/auth/callback/kakao/page.tsx
@@ -1,4 +1,4 @@
-import KakaoLogin from './kakaoLogin';
+import KakaoLogin from '@components/login/KakaoLogin';
 
 export default function KakaoCallback() {
   return (
