+++ src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from './loginPage';
+import LoginPage from '@components/login/MainLogin';
 
 export default function Login() {
   return (
